pyproject.toml
[project]
name = "aula-langchain-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.3.0",
    "langchain-openai>=1.2.1",
    "langgraph>=1.2.0",
    "python-dotenv>=1.2.2",
]
